redid cmake build so unittests are built with the libraries they test. This way unittests are not dependent on the entire code base

+++ rnamake/lib/RNAMake/src/math/numerical.h
@@ -17,9 +17,9 @@
 
 int
 are_floats_equal(
-	float const a,
-	float const b,
-    float tol= 0.001);
+	double const a,
+	double const b,
+    double tol= 0.001);
 
 int
 are_xyzVector_equal(
